requirements.txt
pytelegrambotapi
djangorestframework
djangorestframework-simplejwt
python-decouple
psycopg2-binary
telebot-calendar
celery
redis
